+++ shared/colored_formatter.py
